fixed floating point exception for saving image

+++ train_model.py
@@ -542,13 +542,13 @@ class UNetModel:
                    normalize=True)
 
         for i in range(self.exp_config.num_labels_per_subject):
-            save_image(ground_truth_labels[i],
+            save_image(ground_truth_labels[i].float(),
                        os.path.join(save_location, '{}mask{}.png'.format(iteration, i)),
                        pad_value=1,
                        scale_each=True,
                        normalize=True)
         for i in range(10):
-            save_image(sample[i],
+            save_image(sample[i].float(),
                        os.path.join(save_location, '{}sample{}.png'.format(iteration, i)),
                        pad_value=1,
                        scale_each=True,
